requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
openai==2.29.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
sqlalchemy==2.0.48
jinja2==3.1.6
httpx==0.28.1
fpdf2==2.8.4
markdown==3.9
slowapi==0.1.9
resend==2.7.0
